factor out seat_activation
