+++ services/bundle_analysis/notify/contexts/tests/test_comment_context.py
@@ -249,15 +249,15 @@ class TestBundleAnalysisPRCommentNotificationContext:
             reason="mocked",
         )
         mocker.patch(
-            "services.bundle_analysis.notify.contexts.comment.determine_seat_activation",
+            "services.seats.determine_seat_activation",
             return_value=activation_result,
         )
         mocker.patch(
-            "services.bundle_analysis.notify.contexts.comment.activate_user",
+            "services.seats.activate_user",
             return_value=auto_activate_succeeds,
         )
         mocker.patch(
-            "services.bundle_analysis.notify.contexts.comment.schedule_new_user_activated_task",
+            "services.seats.schedule_new_user_activated_task",
             return_value=auto_activate_succeeds,
         )
         head_commit, _ = get_commit_pair(dbsession)
